+++ utils_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/noop"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
@@ -256,4 +257,84 @@ var omit SpanFilter = func(_ context.Context, _ Method, _ string, _ []driver.Nam
 // keep is a dummy SpanFilter function which specifies to keep the span.
 var keep SpanFilter = func(_ context.Context, _ Method, _ string, _ []driver.NamedValue) bool {
 	return true
+}
+
+func TestRecordMetric(t *testing.T) {
+	type args struct {
+		ctx    context.Context
+		cfg    config
+		method Method
+		query  string
+		args   []driver.NamedValue
+	}
+	tests := []struct {
+		name           string
+		args           args
+		recordErr      error
+		expectedStatus string
+	}{
+		{
+			name: "metric with no error",
+			args: args{
+				cfg:    newConfig(),
+				method: MethodConnQuery,
+				query:  "example query",
+			},
+			recordErr:      nil,
+			expectedStatus: "ok",
+		},
+		{
+			name: "metric with an error",
+			args: args{
+				cfg:    newConfig(),
+				method: MethodConnQuery,
+				query:  "example query",
+			},
+			recordErr:      assert.AnError,
+			expectedStatus: "error",
+		},
+		{
+			name: "metric with skip error but not disabled",
+			args: args{
+				cfg:    newConfig(),
+				method: MethodConnQuery,
+				query:  "example query",
+			},
+			recordErr:      driver.ErrSkip,
+			expectedStatus: "error",
+		},
+		{
+			name: "metric with skip error but disabled",
+			args: args{
+				cfg:    newConfig(WithDisableSkipErrMeasurement(true)),
+				method: MethodConnQuery,
+				query:  "example query",
+			},
+			recordErr:      driver.ErrSkip,
+			expectedStatus: "ok",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			mockLatency := &float64HistogramMock{}
+			mockInstruments := &instruments{
+				latency: mockLatency,
+			}
+			recordFunc := recordMetric(tt.args.ctx, mockInstruments, tt.args.cfg, tt.args.method, tt.args.query, tt.args.args)
+			recordFunc(tt.recordErr)
+			assert.Equal(t, tt.expectedStatus, mockLatency.status)
+		})
+	}
+}
+
+type float64HistogramMock struct {
+	// Add metric.Float64Histogram so we only need to implement the function we care about for the mock
+	metric.Float64Histogram
+	status string
+}
+
+func (m *float64HistogramMock) Record(_ context.Context, _ float64, opts ...metric.RecordOption) {
+	attr := metric.NewRecordConfig(opts).Attributes()
+	statusVal, _ := attr.Value(queryStatusKey)
+	m.status = statusVal.AsString()
 }