+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- `DisableSkipErrMeasurement` option suppresses the recording of `driver.ErrSkip` measurements as errors if enabled. (#389)
+
 ## [0.35.0] - 2024-10-11
 
 ### Changed